requirements.txt
alembic==1.7.4
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
PyJWT==2.3.0
python-dotenv==0.19.1
six==1.16.0
SQLAlchemy==1.4.26
uuid==1.30
Werkzeug==2.0.2
